% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tables.R, R/utils.R
\name{out_of_bounds}
\alias{out_of_bounds}
\title{Identify Out-of-Bounds Values (Outliers)}
\usage{
out_of_bounds(x, na.rm = TRUE)

out_of_bounds(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\value{
A logical vector where TRUE indicates the value is out of bounds.

A logical vector where TRUE indicates the value is out of bounds.
}
\description{
Uses the 1.5 * IQR rule to identify values that are potential outliers.

Uses the 1.5 * IQR rule to identify values that are potential outliers.
}
