% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{theme_rfapa}
\alias{theme_rfapa}
\title{Custom APA Theme for ggplot2}
\usage{
theme_rfapa()
}
\value{
A ggplot2 theme object with APA styling and blank backgrounds.
}
\description{
Applies the APA theme from the `jtools` package and removes panel and plot backgrounds for a cleaner look.
}
\examples{
library(ggplot2)
library(jtools)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_rfapa()
}
